+++ projects/dumbql/vue/src/public-api.ts
@@ -1,5 +1,6 @@
 export { createDumbqlPlugin, useClient, DUMBQL_CLIENT_KEY } from './lib/plugin';
-export { useQuery, type UseQueryResult } from './lib/use-query';
-export { useMutation, type UseMutationResult, type UseMutationFn } from './lib/use-mutation';
-export { useSubscription, type UseSubscriptionResult } from './lib/use-subscription';
+export { useQuery, type UseQueryOptions, type UseQueryResult, type NetworkStatus } from './lib/use-query';
+export { useMutation, type UseMutationOptions, type UseMutationResult, type UseMutationFn } from './lib/use-mutation';
+export { useSubscription, type UseSubscriptionOptions, type UseSubscriptionResult } from './lib/use-subscription';
+export { useLiveQuery, type UseLiveQueryOptions, type UseLiveQueryResult } from './lib/use-live-query';
 export { gql, isSuccess, isError, unwrap, unwrapOrThrow } from '@dumbql/client';